change int

+++ System/Classes/tm_timer.cpp
@@ -15,10 +15,10 @@
 #include "merge_sort.hpp"
 #include "tm_ostream.hpp"
 
-static hashmap<string, int> timing_level (0);
-static hashmap<string, int> timing_nr (0);
-static hashmap<string, int> timing_cumul (0);
-static hashmap<string, int> timing_last (0);
+static hashmap<string, long> timing_level (0);
+static hashmap<string, long> timing_nr (0);
+static hashmap<string, long> timing_cumul (0);
+static hashmap<string, long> timing_last (0);
 
 /******************************************************************************
  * Getting the time
@@ -59,7 +59,7 @@ texmacs_time () {
 void
 bench_start (string task) {
   // start timer for a given type of task
-  if (timing_level[task] == 0) timing_last (task)= (int) texmacs_time ();
+  if (timing_level[task] == 0) timing_last (task)= (long) texmacs_time ();
   timing_level (task)++;
 }
 
@@ -68,7 +68,7 @@ bench_cumul (string task) {
   // end timer for a given type of task, but don't reset timer
   timing_level (task)--;
   if (timing_level[task] == 0) {
-    int ms= ((int) texmacs_time ()) - timing_last (task);
+    long ms= ((long) texmacs_time ()) - timing_last (task);
     timing_nr (task)++;
     timing_cumul (task)+= ms;
     timing_last->reset (task);
@@ -96,7 +96,7 @@ void
 bench_print (tm_ostream& ostream, string task) {
   // print timing for a given type of task
   if (DEBUG_BENCH) {
-    int nr= timing_nr[task];
+    long nr= timing_nr[task];
     ostream << "Task '" << task << "' took " << timing_cumul[task] << " ms";
     if (nr > 1) ostream << " (" << nr << " invocations)";
     ostream << LF;
@@ -104,7 +104,7 @@ bench_print (tm_ostream& ostream, string task) {
 }
 
 static array<string>
-collect (hashmap<string, int> h) {
+collect (hashmap<string, long> h) {
   array<string>    a;
   iterator<string> it= iterate (h);
   while (it->busy ())
